requirements/dev.txt
alembic==1.1.0
Faker==2.0.1
Flask-Assets==0.12
Flask-CLI==0.4.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-RQ==0.2
#Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
honcho==1.0.1
jsmin==3.0 #2.2.2
Mako==1.1.0
psycopg2==2.9.3
#psycopg2-binary==2.9.3
python-dotenv==1.0.0
python-editor==1.0.4
raygun4py==4.3.0
redis==3.3.8
rq==1.1.0
SQLAlchemy==1.3.8
webassets==0.12.1
WTForms==2.2.1
